+++ PhotoGENius/Program.cs
@@ -56,9 +56,10 @@ class Program
             description: "Name of the PFM file to create.",
             getDefaultValue: () => "../Media/imgs_pfm/output.pfm");
             
-        var pngOutput = new Option<string>(
-            name: "--png-output",
-            description: "Name of the PNG file to create.",
+        var output = new Option<string>(
+            name: "--output",
+            description: "Path of the file to create, with extension [Suggestion: ../Media/imgs_png/output.png or" +
+                         "../Media/imgs_jpeg/output.jpeg ].",
             getDefaultValue: () => "../Media/imgs_png/output.png");
             
         var cameraType = new Option<string>(
@@ -109,6 +110,11 @@ class Program
             description: "Number of sample per pixel (must be a perfect square).",
             getDefaultValue: () => 4);
         
+        var format = new Option<string>(
+            name: "--output-format",
+            description: "Output file extension; can be PNG or JPEG",
+            getDefaultValue: () => "PNG");
+        
         //==============================================================================================================
         // Demo 
         //==============================================================================================================
@@ -119,7 +125,7 @@ class Program
             height,
             angleDeg,
             pfmOutput,
-            pngOutput,
+            output,
             cameraType,
             algorithm,
             raysNum,
@@ -128,13 +134,14 @@ class Program
             initSeq,
             luminosityFactor,
             gammaFactor,
-            samplePerPixel
+            samplePerPixel,
+            format
         };
             
         rootCommand.AddCommand(demo);
         demo.SetHandler((int widthValue, int heightValue, float angleDegValue, string pfmOutputValue,
-                string pngOutputValue, string cameraTypeValue, string algorithmValue, int raysNumValue, int maxDepthValue, ulong initStateValue,
-                ulong initSeqValue, float luminosityFactorValue, float gammaFactorValue, int samplePerPixelValue) =>
+                string outputValue, string cameraTypeValue, string algorithmValue, int raysNumValue, int maxDepthValue, ulong initStateValue,
+                ulong initSeqValue, float luminosityFactorValue, float gammaFactorValue, int samplePerPixelValue, string formatValue) =>
             { 
                 var samplePerSide = (int) Math.Sqrt(samplePerPixelValue);
                     
@@ -225,8 +232,10 @@ class Program
                 }
                 else
                 {
-                    throw new Exception("Invalid camera option: use orthogonal, or perspective");
-                }
+                    Console.WriteLine("Invalid camera option: use orthogonal, or perspective");
+                    return;
+                } 
+                
 
                 // 3. Run raytracer
                 Console.WriteLine("\nRunning raytracer...");
@@ -293,30 +302,31 @@ class Program
                         $"Error: couldn't write file {pfmOutputValue}");
                 }
                     
-                // 6. Convert to PNG
-                Console.WriteLine("\nConverting and saving PNG image...");
+                // 6. Convert to LDR
+                Console.WriteLine("\nConverting and saving LDR image...");
                     
                 image.NormalizeImage(luminosityFactorValue);
                 image.ClampImage();
 
-                // save PNG file, depending on options
+                // Save LDR file, depending on options
                 try
                 {
                     {
-                        image.WriteLdrImage(pngOutputValue, "PNG", gammaFactorValue);
+                        image.WriteLdrImage(outputValue, formatValue, gammaFactorValue);
                     }
-
-                    Console.WriteLine($"    File {pngOutputValue} has been written to disk.");
+                    
+                    Console.WriteLine($"    File {outputValue} has been written to disk.");
+                    
                 }
                 catch
                 {
                     Console.WriteLine(
-                        $"Error: couldn't write file {pngOutputValue}");
+                        $"Error: couldn't write file {outputValue}. May be: wrong path, non existent directory, invalid format.");
                 }
                     
             },
-            width, height, angleDeg, pfmOutput, pngOutput, cameraType, algorithm, raysNum, maxDepth, 
-            initState, initSeq, luminosityFactor, gammaFactor , samplePerPixel);
+            width, height, angleDeg, pfmOutput, output, cameraType, algorithm, raysNum, maxDepth, 
+            initState, initSeq, luminosityFactor, gammaFactor , samplePerPixel, format);
             
         //==============================================================================================================
         // Render 
@@ -326,11 +336,7 @@ class Program
             name: "--file-name",
             description: "Input file for scene description",
             getDefaultValue: () => "../InputSceneFiles/DEFAULT_INPUT.txt");
-        var format = new Option<string>(
-            name: "--output-format",
-            description: "Input output file extension; can be PNG or JPEG",
-            getDefaultValue: () => "PNG");
-
+        
         var render = new Command("render", "Create an image.")
         {
             scenefile,
@@ -338,7 +344,7 @@ class Program
             height,
             angleDeg,
             pfmOutput,
-            pngOutput,
+            output,
             algorithm,
             raysNum,
             maxDepth,
@@ -352,7 +358,7 @@ class Program
         rootCommand.AddCommand(render);
             
         render.SetHandler((string scenefileValue, int widthValue, int heightValue, float angleDegValue,
-                string pfmOutputValue, string pngOutputValue, string algorithmValue, int raysNumValue,
+                string pfmOutputValue, string outputValue, string algorithmValue, int raysNumValue,
                 int maxDepthValue, ulong initStateValue, ulong initSeqValue, float luminosityFactorValue,
                 float gammaFactorValue, int samplePerPixelValue, string formatValue) =>
             {
@@ -372,8 +378,16 @@ class Program
                 Stream sceneStream = new FileStream(scenefileValue, FileMode.Open);
                 Dictionary<string, float> dict = new Dictionary<string, float>();
             
-                Scene scene = new Scene();    
-                scene = ExpectParse.parse_scene(new InputStream(sceneStream), dict);
+                Scene scene = new Scene();
+
+                try
+                {
+                    scene = ExpectParse.parse_scene(new InputStream(sceneStream), dict);
+                }
+                catch 
+                {
+                    Console.WriteLine($"Error: could not open {scenefileValue}. \n Try checking your filepath. \n");
+                }
                 
                 // Camera rotation
                 if (angleDegValue != 0)
@@ -381,7 +395,7 @@ class Program
 
                 // 2. Run raytracer
                 Console.WriteLine("\nRunning raytracer...");
-
+                
                 var image = new HdrImage(widthValue, heightValue);
                 Console.WriteLine(
                     $"    Generating a {widthValue}×{heightValue} image, with the camera tilted by {angleDegValue}°");
@@ -447,30 +461,31 @@ class Program
                         $"Error: couldn't write file {pfmOutputValue}");
                 }
                 
-                // 4. Convert to PNG
-                Console.WriteLine("\nConverting and saving PNG image...");
+                // 4. Convert to LDR
+                Console.WriteLine("\nConverting and saving LDR image...");
                 
                 image.NormalizeImage(luminosityFactorValue);
                 image.ClampImage();
 
-                // Save PNG file, depending on options
+                // Save LDR file, depending on options
                 try
                 {
                     {
-                        image.WriteLdrImage(pngOutputValue, formatValue, gammaFactorValue);
+                        image.WriteLdrImage(outputValue, formatValue, gammaFactorValue);
                     }
-
-                    Console.WriteLine($"    File {pngOutputValue} has been written to disk.");
+                    
+                    Console.WriteLine($"    File {outputValue} has been written to disk.");
+                    
                 }
                 catch
                 {
                     Console.WriteLine(
-                        $"Error: couldn't write file {pngOutputValue}. May be: wrong path, inexisting directory, invalid format.");
+                        $"Error: couldn't write file {outputValue}. May be: wrong path, non existent directory, invalid format.");
                 }
             },
 
-            scenefile, width, height, angleDeg, pfmOutput, pngOutput, algorithm, raysNum, maxDepth, 
-            initState, initSeq, luminosityFactor, gammaFactor , samplePerPixel );
+            scenefile, width, height, angleDeg, pfmOutput, output, algorithm, raysNum, maxDepth, 
+            initState, initSeq, luminosityFactor, gammaFactor , samplePerPixel, format );
             
             
         //==============================================================================================================
@@ -481,6 +496,10 @@ class Program
             name: "--pfm-input",
             description: "PFM file to be converted.",
             getDefaultValue: () => "../Media/Readme_imgs/memorial.pfm");
+        var pngOutput = new Option<string>(
+            name: "--png-output",
+            description: "PNG output file name.",
+            getDefaultValue: () => "../Media/imgs_png/output.png");
             
         var pfm2png = new Command("pfm2png", "Convert a PFM file to a PNG")
         {
@@ -528,7 +547,7 @@ class Program
                 catch
                 {
                     Console.WriteLine(
-                        "Error: couldn't write file {0}.", pngOutputValue);
+                        $"Error: couldn't write file {pngOutputValue}. May be: wrong path, non existent directory, invalid format.");
                 }
             },
             luminosityFactor, gammaFactor, pfmInput, pngOutput);